corpus/pyproject.toml
[project]
name = "corpus"
version = "0.1.0"
description = "External content discovery microservice for VectorSpace"
authors = [
    {name = "VectorSpace Team"}
]
readme = "SETUP.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
    "chromadb>=0.4.15",
    "openai>=1.3.0",
    "pydantic>=2.4.0",
    "python-dotenv>=1.0.0",
    "aiolimiter>=1.1.0",
    "schedule>=1.2.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "black>=23.9.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "flake8>=6.0.0",
    "respx>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true